Bus service+controller

+++ src/main/java/com/busticketbooking/repository/TicketRepository.java
@@ -3,7 +3,5 @@ package com.busticketbooking.repository;
 import com.busticketbooking.model.Ticket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TicketRepository extends JpaRepository<Ticket, Integer> {
-
-}
+public interface TicketRepository extends JpaRepository<Ticket, Long> {}
